+++ tests/test_backend.py
@@ -16,7 +16,7 @@ import numpy as np
 import tensorflow as tf
 
 from tensorflow.python.ops import lookup_ops
-from tensorflow.python.ops import init_ops
+
 from backend_test_base import Tf2OnnxBackendTestBase
 # pylint reports unused-wildcard-import which is false positive, __all__ is defined in common
 from common import *  # pylint: disable=wildcard-import,unused-wildcard-import
@@ -2932,27 +2932,6 @@ class BackendTests(Tf2OnnxBackendTestBase):
 
         for input_val in input_vals:
             self._run_test_case(func, [_OUTPUT], {_INPUT: input_val})
-
-    @skip_tf_cpu("only tf_gpu can run CudnnGPU")
-    @check_opset_min_version(11, "GRU")
-    def test_cudnngru(self):
-        def func():
-            seq_length = 3
-            batch_size = 5
-            input_size = 2
-            num_layers = 2
-            num_units = 2
-            num_dirs = 2
-            initializer = init_ops.constant_initializer(0.5)
-            x = np.random.randint(0, 100, [seq_length, batch_size, input_size]).astype(np.float32)
-            h = np.random.randint(0, 100, [num_layers * num_dirs, batch_size, num_units]).astype(np.float32).reshape(
-                [num_layers * num_dirs, batch_size, num_units])
-            cudnngru = tf.contrib.cudnn_rnn.CudnnGRU(num_layers, num_units, 'linear_input', 'bidirectional',
-                                                     kernel_initializer=initializer, bias_initializer=initializer)
-            cudnngru.build([seq_length, batch_size, input_size])
-            outputs = cudnngru.call(x, tuple([h]))
-            _ = tf.identity(outputs[0], name=_TFOUTPUT)
-        self.run_test_case(func, {}, [], [_OUTPUT], rtol=1e-05, atol=1e-04)
 
 
 if __name__ == '__main__':
